created apiRoutes and js code, utility functions, edited pug template for projects

+++ routes/utils.js
@@ -1,15 +1,19 @@
-const asyncHandler = (handler) => (req, res, next) => handler(req, res, next).catch(next)
-const csrf = require('csurf')
-const csrfProtection = csrf({ cookie: true })
-
-
-
+const asyncHandler = (handler) => (req, res, next) => handler(req, res, next).catch(next);
+const csrf = require("csurf");
+const csrfProtection = csrf({ cookie: true });
 
 //project delete
 const deleteItem = async (itemId, model) => {
-  const item = await model.findByPk(itemId)
-  await item.destroy()
-}
+	const item = await model.findByPk(itemId);
+	await item.destroy();
+};
 
+// QUERY FUNCTIONS
+const findCurrentUser = (session) => {
+	const {
+		userAuth: { userId },
+	} = session;
+	return userId;
+};
 
-module.exports = { asyncHandler, csrfProtection, deleteItem }
+module.exports = { asyncHandler, csrfProtection, deleteItem, findCurrentUser };
